# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: POS + Inventory + Delivery Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transfers:
    get:
      operationId: getTransfers
      tags: [transfers]
      summary: List transfer history
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
    post:
      operationId: createTransfer
      tags: [transfers]
      summary: Transfer items between invoices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferInput"
      responses:
        "201":
          description: Transfer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResult"

  # ── Reports ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: number
        price:
          type: number
        subtotal:
          type: number
        isDamaged:
          type: boolean
      required: [id, invoiceId, productId, productName, qty, price, subtotal, isDamaged]
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNo:
          type: string
        customerName:
          type: string
        date:
          type: string
        total:
          type: number
        deliveryId:
          type: integer
          nullable: true
        deliveryNo:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
      required: [id, invoiceNo, customerName, date, total]
    InvoiceWithItems:
      allOf:
        - $ref: "#/components/schemas/Invoice"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/InvoiceItem"
          required: [items]
...
    Transfer:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: number
        fromInvoiceId:
          type: integer
        fromInvoiceNo:
          type: string
        fromCustomer:
          type: string
        toInvoiceId:
          type: integer
        toInvoiceNo:
          type: string
        toCustomer:
          type: string
        date:
          type: string
        note:
          type: string
          nullable: true
      required: [id, productId, productName, qty, fromInvoiceId, fromInvoiceNo, fromCustomer, toInvoiceId, toInvoiceNo, toCustomer, date]
    CreateTransferInput:
      type: object
      properties:
        fromInvoiceId:
          type: integer
        toInvoiceId:
          type: integer
          nullable: true
        toCustomerName:
          type: string
          nullable: true
        productId:
          type: integer
        qty:
          type: number
        note:
          type: string
          nullable: true
      required: [fromInvoiceId, productId, qty]
    TransferResult:
      type: object
      properties:
        transfer:
          $ref: "#/components/schemas/Transfer"
        updatedFromInvoice:
          $ref: "#/components/schemas/InvoiceWithItems"
        newOrUpdatedToInvoice:
          $ref: "#/components/schemas/InvoiceWithItems"
      required: [transfer, updatedFromInvoice, newOrUpdatedToInvoice]
...